requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cursor-pagination==0.2.1
django-extensions==3.2.1
idna==3.4
Pillow==9.3.0
pkg_resources==0.0.0
python-decouple==3.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
